// types/user.types.ts
// user.types.ts

export interface User {
    id: number;
    name: string;
    username: string;
    email: string;
    job: string;
    gender: string;
    phone_number: string;
    website: string;
    address: string;
    birthdate: string;
    age: number;
    interests: string[];
    user_image_url: string;
    background_image_url: string;
}